requirements.txt
requests~=2.31.0
python-dotenv~=1.0.1
infi.clickhouse_orm~=2.1.3